If CodeRay parsing fails, fallback to just escape entities.
When lang is not set, because there is not a specific language to highlight
for, coderay will fail silently. Previously this lead to the code to never
be escaped, causing the lesser/greater characters to be interpreted as
(X)HTML code.

When CodeRay fails, fallback on using HTMLEntities for the encoding, which
does not care about the language at all.