Only do page caching if there's no query string.
This solves the problem where the first time a page is accessed, the client
passes additional parameters in the query string that change the way the
page looks, causing the wrong version to be presented to later clients.

The first access without query string is the one that gets cached. After
that, query strings have no effect on what gets shown anymore, but that's
ok, since the client is not supposed to add query strings to these URLs
anyway.