Fix links to images in administration style sheets.
The links went up two levels, which is one level too many. This is no
problem if Typo is installed at the top level, because the extra level will
just be ignored. When Typo is installed in a sub-URI, it will cause all the
UI images to be missing. The links now go up just one level.