When finding articles by permalink, use UTC date.
Up to typo 5.5, the dates as found in the article permalink URLs
corresponded exactly with the publication date as stored in the database,
and these dates were in the local timezone.

Starting with Rails 3 however, datetimes as stored in the database are
interpreted as having zone UTC, breaking existing URLs and causing a
mismatch between the generated URLs (which due to how they are programmed
were immune to this change) and the permalink finder method.

This commit restores the old behavior, i.e., the permalink dates correspond
again to the dates as stored on disk.

However, the dates continue to be interpreted as UTC, which means the
publication times for old articles (originally stored as local time) have
an offset compared to the publication times for new articles (stored as
UTC).

Some tests had to be adjusted to be more specific about the generated
publication datetime.