Migrating users to the same config manager we're using for blogs.
The goal is to be able to dynamically extend users with new settings without adding a new collumn each time.

The migration is a bit ugly.

Problem: settings names overlap fields names.
Solution: rename columns, migrate, remove columns.

There must be a better way to do it but...
1. I didn't find it.
2. I'm lost in the countryside where I need to go to the back of the
garden to have a phone connection and in the middle of a field to have
 a chance to get my emails. So asking for help / advice on the interweb was not possible.

At least it works

Some fields were not migrated because we do some User.find on them :
– notify_via_email
– notify_on_new_articles
– notify_on_comments