Create a second migration to fix article post types.
Because ActiveRecord::Base.store_full_sti_class is now true by default
(since Rails 3.0.0), migration 102 was looking for Content instances where
type was equal to "FixesExistingArticlesWithNoType::Article", causing no
records to be found. Since this migration is in a release, we really cannot
just fix it. So, we add a second migration that proposes to do the same
thing.