Fixes switching back and fort between simple and visual editor.
When switching back and fort from visual editor to simple editor, data would still appear in the ckeditor but was never sent to server. It seems that using CKEDITOR.instances.page__body_editor.setData() was not enough as the page__body_editor hidden textarea had to be filled in first.