Remove useless calls to #changed.
The intent of this code seems to have been to avoid model objects being
marked dirty when the new value is the same as the old. However, this code
didn't do that since #changed just returns the names of the changed
attributes. Also, current ActiveRecord already avoids marking models as
dirty when the new value is the same as the old.