Removing theme catalogue and mentions of Typogarden.
Theme catalogue now returns a 404 and is not maintained anymore.
We've lost typogarden.org (I actually forgot to renew it) and most themes were not supported anymore.

We'll need to find a way to display themes, maybe using themes.typosphere.org with a statically cached blog, I need to think about this first.