Remove the Rights model.
A long, long time ago, I wanted Typo permission systems to be as flexible as possible (and as user friendly as possible). My idea was to make the various profiles editable with a set of rights, everything within a fancy interface.

I started to push some things, then completely forgot about it. As this makes our permission system way more complicated than it should be, without a real interest, I'm now removing it all.

See #79 for more information.