Fix live_search by outputting the ERB form block
The live_search sidebar module is missing the search field.  Looking at
the markup, the surrounding form element is missing as well.  Newer
Rails versions require that blocks in ERB use the "<%=" output delimiter
to output markup generated within the block, but the live_search module
is not using that.
    
Switch to the output-style delimiter for the form_tag block.