Show fallback account name on header bar when logged in admin
By default the admin header displays (where X is the nickname)

  "Logged in as X"

If the account does not have a nickname set it will be blank
and instead just show

  "Logged in as "

Do not assume nickname has been chosen yet and instead fallback
to either name, or if that is blank, login username. The last
is never blank.