Fix latest posts date format with distance_of_time_in_words
The Latest Posts on the admin dashboard is showing
"distance_of_time_in_words" as the date of each post, because that
is what is configured as my blog's date format.  Elsewhere, this
date format is treated as a special case rather than a strftime format.

Format the date with display_date_and_time rather than directly calling
strftime on the date and time format of the blog.