Add background highlight to currently active theme in admin
Instead of a small bit of italic text, mark out the current theme
in use with an alert DIV. Clearer, quick to see if you have many
themes to choose from.

Tidy up the if-else block uses as the contents are mostly different
between inactive choices and the active one.

"Use this theme" sounded more direct and action packed than
"Choose this theme" for the button.

Cleans up language file definitions orphaned by removal of
references to Typogarden and installing themes in
commit 82c73cd62c3fc616cb44660d9ac6b4dfbe53ddb9