Fixes feedback bulkops for mark as spam that would actually mark as ham.
Culprit is merge 3bdbbbec28a03c7d1e6b12adfb4bd77854702057

If

  def mark_as_spam!
    mark_as_ham
    save!
  end

doesn't strike you...

The specs were not falling becaus for some reasons, comment.reload did not reload the comment state. As the test was done with a presumably spammy comment that was marked as spam, the status stayed the same and specs did not crash.

This commit fixes the issue and adds specs for all bulkops cases.